requirements.txt
tornado>=6.0
streamlit
streamlit-webrtc
opencv-python
pyaudio
protobuf==5.26.1
setuptools==70.0.0
grpcio==1.67.1
grpcio-tools==1.67.1
matplotlib
librosa
mediapipe
python-dotenv
requests
supabase


langchain_chroma
langchain
langchain_openai
langchain_text_splitters
langchain_community
pypdf
faiss-cpu
ploty